Refactor: 인증, 인가 기능 구현

+++ src/main/java/trendravel/photoravel_be/domain/photographer/service/PhotographerService.java
@@ -2,20 +2,32 @@ package trendravel.photoravel_be.domain.photographer.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import trendravel.photoravel_be.commom.error.MemberErrorCode;
 import trendravel.photoravel_be.commom.error.PhotographerErrorCode;
 import trendravel.photoravel_be.commom.exception.ApiException;
 import trendravel.photoravel_be.commom.image.service.ImageServiceFacade;
 import trendravel.photoravel_be.db.enums.Region;
+import trendravel.photoravel_be.db.inmemorydb.entity.Token;
 import trendravel.photoravel_be.db.photographer.Photographer;
 import trendravel.photoravel_be.db.respository.photographer.PhotographerRepository;
 import trendravel.photoravel_be.db.review.Review;
+import trendravel.photoravel_be.domain.authentication.service.PhotographerAuthenticationService;
+import trendravel.photoravel_be.domain.authentication.session.PhotographerSession;
 import trendravel.photoravel_be.domain.photographer.dto.request.PhotographerRequestDto;
 import trendravel.photoravel_be.domain.photographer.dto.request.PhotographerUpdateDto;
 import trendravel.photoravel_be.domain.photographer.dto.response.PhotographerListResponseDto;
 import trendravel.photoravel_be.domain.photographer.dto.response.PhotographerSingleResponseDto;
 import trendravel.photoravel_be.domain.review.dto.response.RecentReviewsDto;
+import trendravel.photoravel_be.domain.token.model.TokenDto;
+import trendravel.photoravel_be.domain.token.model.TokenResponse;
+import trendravel.photoravel_be.domain.token.service.TokenService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,6 +39,10 @@ public class PhotographerService {
     
     private final PhotographerRepository photographerRepository;
     private final ImageServiceFacade imageServiceFacade;
+    private final PasswordEncoder photographerBCryptPasswordEncoder;
+    private final TokenService tokenService;
+    private final RedisTemplate<String, Token> redisTemplate;
+    private final PhotographerAuthenticationService photographerAuthenticationService;
     
     @Transactional
     public List<PhotographerListResponseDto> getPhotographerList(String region) {
@@ -96,7 +112,7 @@ public class PhotographerService {
         
         photographerRepository.save(Photographer.builder()
                 .accountId(photographerRequestDto.getAccountId())
-                .password(photographerRequestDto.getPassword())
+                .password(photographerBCryptPasswordEncoder.encode(photographerRequestDto.getPassword()))
                 .name(photographerRequestDto.getName())
                 .region(photographerRequestDto.getRegion())
                 .description(photographerRequestDto.getDescription())
@@ -108,31 +124,51 @@ public class PhotographerService {
     }
     
     @Transactional
-    public void authenticate(String username, String password) {
+    public TokenResponse authenticate(String username, String password) {
         
         Photographer photographer = photographerRepository.findByAccountId(username).orElseThrow(() ->
                 new ApiException(PhotographerErrorCode.PHOTOGRAPHER_NOT_FOUND));
-        
-        //인증 로직 필요
-        
+
+        if (photographerBCryptPasswordEncoder.matches(password, photographer.getPassword())) {
+            return issueTokenResponse(photographer);
+        } else {
+            throw new ApiException(MemberErrorCode.PASSWORD_NOT_MATCH);
+        }
+
     }
     
     @Transactional
     public PhotographerSingleResponseDto updatePhotographer(PhotographerUpdateDto photographerUpdateDto,
                                                             List<MultipartFile> images) {
-        
-        Photographer photographer = photographerRepository.findByAccountId(photographerUpdateDto.getAccountId()).orElseThrow(
-                () -> new ApiException(PhotographerErrorCode.PHOTOGRAPHER_NOT_FOUND));
-        
+
+        PhotographerSession principal = (PhotographerSession) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Photographer photographer = photographerRepository.findByAccountId(principal.getUsername())
+                .orElseThrow(() -> new ApiException(PhotographerErrorCode.UNAUTHORIZED));
+
+        redisTemplate.opsForHash().delete("photographer_refresh_token", photographer.getAccountId());
+
         //기존 이미지 삭제
         List<String> originImage = new ArrayList<>();
         originImage.add(photographer.getProfileImg());
         imageServiceFacade.deleteAllImagesFacade(originImage);
-        
+
         photographer.updatePhotographer(photographerUpdateDto, imageServiceFacade.uploadImageFacade(images));
-        
+
         List<RecentReviewsDto> reviews = photographerRepository.recentReviews(photographer.getId());
-        
+
+        TokenDto accessToken = tokenService.issuePhotographerAccessToken(photographer.getAccountId());
+        TokenDto refreshToken = tokenService.issuePhotographerRefreshToken(photographer.getAccountId());
+
+        TokenResponse tokenResponse = TokenResponse.builder()
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .build();
+
+        // 사진작가 정보 수정 후 UserDetails를 업데이트
+        PhotographerSession photographerSession = (PhotographerSession) photographerAuthenticationService.loadUserByUsername(photographerUpdateDto.getAccountId());
+        Authentication authentication = new UsernamePasswordAuthenticationToken(photographerSession, null, photographerSession.getAuthorities());
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+
         return PhotographerSingleResponseDto.builder()
                 .accountId(photographer.getAccountId())
                 .name(photographer.getName())
@@ -144,20 +180,36 @@ public class PhotographerService {
                 .ratingAvg(String.format("%.2f", ratingAverage(photographer.getReviews())))
                 .recentReviewDtos(reviews)
                 .careerYear(photographer.getCareerYear())
+                .tokenResponse(tokenResponse)
                 .build();
     }
     
     @Transactional
     public PhotographerSingleResponseDto updatePhotographer(PhotographerUpdateDto photographerUpdateDto) {
-        
-        Photographer photographer = photographerRepository.findByAccountId(photographerUpdateDto.getAccountId()).orElseThrow(
-                () -> new ApiException(PhotographerErrorCode.PHOTOGRAPHER_NOT_FOUND));
-        
+
+        PhotographerSession principal = (PhotographerSession) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Photographer photographer = photographerRepository.findByAccountId(principal.getUsername())
+                .orElseThrow(() -> new ApiException(PhotographerErrorCode.UNAUTHORIZED));
+        redisTemplate.opsForHash().delete("photographer_refresh_token", photographer.getAccountId());
+
         photographer.updatePhotographer(photographerUpdateDto);
-        
         List<RecentReviewsDto> reviews = photographerRepository.recentReviews(photographer.getId());
-        
+
+        TokenDto accessToken = tokenService.issuePhotographerAccessToken(photographer.getAccountId());
+        TokenDto refreshToken = tokenService.issuePhotographerRefreshToken(photographer.getAccountId());
+
+        TokenResponse tokenResponse = TokenResponse.builder()
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .build();
+
+        // 사진작가 정보 수정 후 UserDetails를 업데이트
+        PhotographerSession photographerSession = (PhotographerSession) photographerAuthenticationService.loadUserByUsername(photographer.getAccountId());
+        Authentication authentication = new UsernamePasswordAuthenticationToken(photographerSession, null, photographerSession.getAuthorities());
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+
         return PhotographerSingleResponseDto.builder()
+                // id는 만들어 놓고 안 넣으신 이유가 무엇인지..?
                 .accountId(photographer.getAccountId())
                 .name(photographer.getName())
                 .region(photographer.getRegion())
@@ -168,20 +220,30 @@ public class PhotographerService {
                 .ratingAvg(String.format("%.2f", ratingAverage(photographer.getReviews())))
                 .recentReviewDtos(reviews)
                 .careerYear(photographer.getCareerYear())
+                .tokenResponse(tokenResponse)
                 .build();
     }
     
     @Transactional
     public void deletePhotographer(String photographerId) {
-        
-        Photographer photographer = photographerRepository.findByAccountId(photographerId).orElseThrow(
-                () -> new ApiException(PhotographerErrorCode.PHOTOGRAPHER_NOT_FOUND));
-        
-        List<String> img = new ArrayList<>();
-        img.add(photographer.getProfileImg());
-        imageServiceFacade.deleteAllImagesFacade(img);
-        
-        photographerRepository.deleteByAccountId(photographerId);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        PhotographerSession principal = (PhotographerSession) authentication.getPrincipal();
+
+        Photographer photographer = photographerRepository.findByAccountId(principal.getUsername())
+                .orElseThrow(() -> new ApiException(MemberErrorCode.USER_NOT_FOUND));
+
+        if (photographer.getAccountId().equals(photographerId)){
+            List<String> img = new ArrayList<>();
+            img.add(photographer.getProfileImg());
+            imageServiceFacade.deleteAllImagesFacade(img);
+            photographerRepository.deleteByAccountId(photographerId);
+            redisTemplate.opsForHash().delete("photographer_refresh_token", photographerId);
+            SecurityContextHolder.clearContext();
+        }else {
+            throw new ApiException(MemberErrorCode.UNAUTHORIZED);
+        }
+
         
         /*
         단일 이미지 삭제 로직 구현 필요
@@ -198,6 +260,16 @@ public class PhotographerService {
             sum += review.getRating();
         }
         return sum / reviews.size();
+    }
+
+    private TokenResponse issueTokenResponse(Photographer photographer) {
+        TokenDto accessTokenDto = tokenService.issuePhotographerAccessToken(photographer.getAccountId());
+        TokenDto refreshTokenDto = tokenService.issuePhotographerRefreshToken(photographer.getAccountId());
+
+        return TokenResponse.builder()
+                .accessToken(accessTokenDto)
+                .refreshToken(refreshTokenDto)
+                .build();
     }
